+++ fileio.h
@@ -10,9 +10,9 @@
 #include <sys/types.h>
 
 FILE *fs_open(char *path, uint32_t mode);
-int fs_seek(FILE *stream, off_t offset, int whence);
-off_t fs_read(void *ptr, uint32_t size, uint32_t n, FILE *stream);
-off_t fs_write(const void *ptr, uint32_t size, uint32_t n, FILE *stream);
+int fs_seek(FILE *stream, long offset, int whence);
+off_t fs_read(void *ptr, size_t size, size_t n, FILE *stream);
+off_t fs_write(const void *ptr, size_t size, size_t n, FILE *stream);
 int fs_close(FILE *file);
 
 #endif // __fileio_h__
